si dupcheck bugfix + update readme
