+++ account_supplier_invoice_check_duplicates/models/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+from . import account_invoice
