+++ account_supplier_invoice_check_duplicates/views/account_invoice.xml
@@ -2,8 +2,8 @@
 <openerp>
   <data>
 
-    <record id="invoice_supplier_form_inherit" model="ir.ui.view">
-      <field name="name">account.invoice.supplier.form.inherit</field>
+    <record id="invoice_supplier_form" model="ir.ui.view">
+      <field name="name">account.invoice.supplier.form</field>
       <field name="model">account.invoice</field>
       <field name="inherit_id" ref="account.invoice_supplier_form"/>
       <field name="arch" type="xml">
